+++ .eslintignore
@@ -1,2 +1,4 @@
 webpack.dev.config.js
 webpack.config.js
+.prettierrc
+dist
