+++ .gitignore
@@ -1,3 +1,7 @@
+# Ignoring client files
+/public/styles/app.css
+/public/bundle.js
+
 # Logs
 logs
 *.log
