+++ README.md
@@ -1,17 +1,28 @@
-# projnotes-2023a-class
-Class proyect from the FullStack web development
-SubJect.
+# Projnotes 2023 Edition (Class)
+Class project from the FullStack Web Development 
+Subject.
 
 ## About...
-**Autor:** [Brayan Garcia]()
----
-## Proyect Stack
-This Proyect will be coded in ES6, as runtine
-enviroment the project will use node, for the databases the proyect will use MongoDb Technology, for template engine we will use Handlebars
+**Author:** [Ivan Rivalcoba](https://about.me/ivanrivalcoba/getstarted)
 
-## Proyect Notes
-1.[Proyect Creation](https://github.com/bryangc24/projnotes-2023a-class/tree/main)
-5.
-6.Build Script
-7.babel plugins
-8.Addin a bundler
+---
+
+## Project Stack
+This project will be developed using ECMAScript 6 (ES6) and run on a Node.js runtime environment. MongoDB will be utilized for database management, and Handlebars will be employed as the template engine.
+
+## Project Notes
+1. [Project Creation](https://github.com/rivalcoba/projnotes-2023a-class/blob/main/class-notes/1-Project-Creation.md)
+
+2. Repository Creation
+
+3. Creating dedicated Backend directory
+
+4. Adding Babel
+
+5. Updating to ES6
+
+6. Build Script
+
+7. Babel plugins
+
+8. Adding a bundler
