--- Server/domains/home/home.router.js
@@ -1,17 +1,0 @@
-// Importando el Router de Express
-import { Router } from 'express';
-
-// Importando el controlador
-import homeController from './home.controller';
-
-// Creando una isntancia del enrutador
-const router = new Router();
-
-// Enrutamos
-// GET '/'
-// GET '/home'
-// GET '/index
-router.get(['/', '/home', '/index'], homeController.home);
-
-// Exporto este tramo de ruta
-export default router;
