--- Server/views/author.hbs
@@ -1,9 +1,0 @@
-<h1>Datos de autor</h1>
-<h2>nombre</h2>
-<p> {{ name }} </p>
-<h2>Apellido</h2>
-<p>{{ lastname }}</p>
-<h2>Lugar de trabajo</h2>
-<p>{{ job }}</p>
-<h2>💎Facebook</h2>
-<p>{{ Facebook }}</p>
