--- Server/views/index.hbs
@@ -1,3 +1,0 @@
-<h1>
-   {{icon}} {{title}}</h1>
-<p>Welcome to my project page {{title}}</p>
