--- Server/views/partials/_navbar.hbs
@@ -1,12 +1,0 @@
-<nav>
-  <div class="container">
-    <div class="nav-wrapper">
-      <a href="#" class="brand-logo">LOGO</a>
-      <ul id="nav-mobile" class="right hide-on-med-and-down">
-        <li><a href="#">Oper 1</a></li>
-        <li><a href="#">Oper 2</a></li>
-        <li><a href="#">Oper 3</a></li>
-      </ul>
-    </div>
-  </div>
-</nav>
