+++ client/styles/style.css
@@ -1,13 +1,13 @@
 body{
-    padding: 50px;
-    font-size: 14px;
-    font-family: "Lucida Grande", Helvetica, Arial, sans-serif;
-  }
-  
-  a{
-    color: #00B7FF;
-  }
-  
-  h1{
-    color: teal;
-  }
+  padding: 50px;
+  font-size: 14px;
+  font-family: "Lucida Grande", Helvetica, Arial, sans-serif;
+}
+
+a{
+  color: #00B7FF;
+}
+
+h1{
+  color: teal;
+}
