+++ server/views/author.hbs
@@ -1,0 +1,9 @@
+<h1>Datos del autor</h1>
+<h2>Nombre</h2>
+<p> {{ name }} </p>
+<h2>Apellido</h2>
+<p> {{ lastname }}</p>
+<h2>Lugar de trabajo</h2>
+<p> {{ job }}</p>
+<h2>🐦 Twitter</h2>
+<p> {{ twitter }}</p>
