+++ server/views/index.hbs
@@ -1,0 +1,2 @@
+<h1>{{icon}} {{title}}</h1>
+<p>Welcome to my awesome project page</p>
