+++ server/views/layouts/main.hbs
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>{{title}}</title>
+    <link rel="shortcut icon" href="/favicon.png" type="image/png">
+    <link rel='stylesheet' href='/styles/app.css' />
+    <script src="/bundle.js" defer></script>
+  </head>
+  <body>
+    {{!-- Llamando a la vista parcial de la barra de navegacion --}}
+    {{> _navbar }}
+
+    <div class="container">
+      {{{body}}}
+    </div>
+  </body>
+</html>
