+++ webpack.dev.config.js
@@ -4,9 +4,10 @@
 
 // Importing an file routing manager
 const path = require('path');
-// Importing Extract Plugin
+// Importing plugin
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const EslintPlugin = require("eslint-webpack-plugin");
+const ESlintPlugin = require("eslint-webpack-plugin");
+
 // We export a configuration object
 // that will be used by webpack
 module.exports = {
@@ -22,7 +23,9 @@ module.exports = {
     // of the project's static files
     path: path.resolve(__dirname, "public"),
     // 2.2 Output file name
-    filename: "bundle.js"
+    filename: "bundle.js",
+    // 2.3
+    publicPath: '/'
   },
   // 3. Configuring the development server
   // The development server serves the packaged files
@@ -31,15 +34,18 @@ module.exports = {
     // 3.1 Static files folder
     static: path.join(__dirname, "public"),
     // 3.2 Development server port
-    port: 3000,
+    port: 8080,
     // 3.3 Defining the host
     host: "0.0.0.0"
   },
-   // Adding a module to webpack
-   module: {
+  // Adding a module to webpack
+  module: {
     rules: [
       {
+        // This section stablishes 
+				// what rules to apply to ".js" files
         test: /\.js$/,
+        // We Dont want to transpile any kind of modules
         exclude: /(node_modules|bower_components)/,
         use: [
           {
@@ -51,7 +57,7 @@ module.exports = {
                   {
                     'modules': false,
                     'useBuiltIns': 'usage',
-                    'targets': {"chrome": "80"},//'> 0.25%, not dead',
+                    'targets': {"chrome": "80"},
                     'corejs': 3
                   }
                 ]
@@ -68,10 +74,9 @@ module.exports = {
   },
   plugins: [
     new MiniCssExtractPlugin({
-    // Archivo css de salida
-    filename: 'styles/app.css'
-  }),
-  new EslintPlugin()
-]
+      // Archivo css de salida
+      filename: 'styles/app.css'
+    }),
+    new ESlintPlugin()
+  ]
 }
-
